move each tab-pane into a own class
